client-telegram-bot/pyproject.toml
[project]
name = "client-telegram-bot"
version = "0.1.0"
description = "Telegram Bot client for Recipe Tracker"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "websockets>=13.0",
    "pydantic-settings>=2.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["telegram_bot"]

[project.scripts]
telegram-bot = "telegram_bot.bot:main"
